Agent: Cookie is now a Long

+++ nsdb/src/main/scala/org/midonet/cluster/flowhistory/FlowRecordEncoding.scala
@@ -135,7 +135,7 @@ case class FlowRecordMatch(inputPortNo: Int, tunnelKey: Long,
 
 case class FlowRecord(host: UUID, inPort: UUID,
                       flowMatch: FlowRecordMatch,
-                      cookie: Int,
+                      cookie: Long,
                       devices: List[TraversedDevice],
                       rules: List[TraversedRule],
                       simResult: SimulationResult.SimulationResult,
@@ -157,7 +157,7 @@ object FlowRecord {
                                    GENERATED_PACKET)
         Collections.shuffle(simRes)
         FlowRecord(UUID.randomUUID, UUID.randomUUID, randomMatch(r),
-                   r.nextInt(), randomDevices(r),
+                   r.nextLong(), randomDevices(r),
                    randomRules(r), simRes.get(0),
                    randomUUIDS(r), randomActions(r))
     }
